backend/requirements.txt
fastapi
uvicorn
requests
newsapi-python
spacy
sqlmodel
anthropic
python-dotenv
